Seed metadata with empty breadcrumb during discover phase
There is a failure to select streams in tap-pipedrive v1.0.6.
This is due to the discover phase not seeding the stream metadata
with an empty breadcrumb. When tap-pipedrive is run, it looks for
a 'selected' key in the empty breadcrumb which does not exist.
There is probably a better way of fixing this but I am applying
this change now.